Update style.css
LINE 202: while watching the tutorial, which I really enjoyed, I was having an issue giving calling my socials-container and giving them a margin from the top. I found out that a hashtag was used to call the class instead of period. at least this was the bug in my code. TYVM!